+++ tests/fast_ostream_test.cpp
@@ -57,10 +57,10 @@ BOOST_AUTO_TEST_SUITE( test_fast_ostream )
 BOOST_AUTO_TEST_CASE(int_test)
 {
 
-  BOOST_CHECK(encode_integer<int8_t>(2, false, "\x82") );
-  BOOST_CHECK(encode_integer<int8_t>(2, true, "\x83") );
-  BOOST_CHECK(encode_integer<int8_t>(-2, false, "\xFE") );
-  BOOST_CHECK(encode_integer<int8_t>(-2, true, "\xFE") );
+  BOOST_CHECK(encode_integer<int16_t>(2, false, "\x82") );
+  BOOST_CHECK(encode_integer<int16_t>(2, true, "\x83") );
+  BOOST_CHECK(encode_integer<int16_t>(-2, false, "\xFE") );
+  BOOST_CHECK(encode_integer<int16_t>(-2, true, "\xFE") );
 
   BOOST_CHECK(encode_integer(INT32_C(942755), true, "\x39\x45\xa4" ));
   BOOST_CHECK(encode_integer(INT32_C(942755), false, "\x39\x45\xa3"));
